+++ codegen/_cli.py
@@ -12,11 +12,24 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""
-mad-codegen --lang <lang>...
+"""Meraki API Docs Code Generator
+
+USAGE:
+    mad-codegen [--lang <lang>...] [--spec <spec>...] [--option <option>...]
 
 OPTIONS:
-  -l, --lang <lang>     Output language. Can be specified multiple times.
+  -l, --lang <lang>...      Language to output. Can be specified multiple times
+                            Available: []
+  -s, --spec <spec>...      Spec to output. Can be specified multiple times.
+                            Available: [openapi3, postman]
+  -o, --option <option>...  Options for language or postman generation.
+
+
+DESCRIPTION:
+    Create an API client in your $language.
+    By default, will save the spec or module in the current working directory.
+
+    If no options are specified, default is to output openapi3 json and quit.
 
 SEE ALSO:
   OpenAPI Generator: https://github.com/OpenAPITools/openapi-generator
@@ -26,4 +39,9 @@ import docopt
 
 def get_cli_args():
     """CLI entry point"""
-    return docopt.docopt(__doc__)
+    args = docopt.docopt(__doc__)
+    no_options_specified = not args['--lang'] and not args['--spec']
+    if no_options_specified:
+        args['--spec'] = 'openapi3'
+
+    return args