+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-* OpenAPI to Postman converter
 * Tests
+
+## [0.1.0] - 2019-02-18
+* OpenAPI to Postman converter working
+* CLI vertical now works 
 
 ## [0.0.1] - 2019-02-18
 ### Fixed
@@ -32,5 +35,7 @@ Fixed:      for any bug fixes.
 Security:   for vulnerability fixes.
 -->
 
-[Unreleased]: https://github.com/pocc/apigen/compare/v0.0.0...HEAD
+[Unreleased]: https://github.com/pocc/apigen/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/pocc/apigen/compare/v0.1.0...v0.0.1
+[0.0.1]: https://github.com/pocc/apigen/compare/v0.0.0...v0.0.1
 [0.0.0]: https://github.com/pocc/apigen/commit/ba8dec9